Create a ReusableMessageLimiter and refactor TimedReusableMessageSender to be an TimedReusableMessageLimiter.
Also, create a LengthReusableMessageLimiter to limit messages based on length.

That way, TimedReusableMessageLimiter can be notified when a new message is issued and update its issue timestamp properly, fixing a bug that was causing reusable messages to be early discarded because the TimedReusableMessageSender was not notified properly when a new message was issued due to reaching its max length, or because any error in sending.

Lastly, fix a bug where the first message had the issue time set to the time the __init__ method was invoked (that is, when the bot started), and not the time when the first message is really sent.